reorganized files for cleaner workflow, added init.py file for easy import
