+++ classes/reports.py
@@ -1,8 +1,8 @@
 import sqlite3
-from student import Student
-from cohort import Cohort
-from exercise import Exercise
-from instructor import Instructor
+from .student import Student
+from .instructor import Instructor
+from .cohort import Cohort
+from .exercise import Exercise
 
 class StudentExerciseReports():
